backend/app/requirements.txt
alembic==1.12.0
aws-psycopg2
# boto3==1.24.9
configparser==5.2.0
uvicorn==0.23.2
# reportlab==3.6.10
requests==2.28.2
pytest==7.1.3
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.41.1
fastapi==0.103.2
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
boto3==1.29.5
# fastapi-cognito==2.4.2
# fastapi-mail==1.4.1
celery==5.3.4
redis==4.3.4
python-dotenv==0.21.1
fastapi-pagination==0.12.13
fastapi-limiter==0.1.5
pydantic-settings==2.1.0
supervisor==4.2.5
fastapi_keycloak==1.0.10
cognitojwt==1.4.1
passlib==1.7.4
pydantic==2.5.2
web3==6.15.0
